allow for easy transition of blog posts
