+++ src/js/components/blog/blog_article/articleBody.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ArticleBody = ({}) => {
+
+  // return ( )
+
+}
